principal structure categories of css
