+++ docs/containers/Routes.js
@@ -47,14 +47,6 @@ const Routes = () =>
         <Route
           getComponent={(_nextState, callback) => {
             require.ensure([], require => {
-              callback(null, require('./css/Base'))
-            }, 'base')
-          }}
-          path="base"
-        />
-        <Route
-          getComponent={(_nextState, callback) => {
-            require.ensure([], require => {
               callback(null, require('./css/Tools'))
             }, 'tools')
           }}
@@ -91,6 +83,14 @@ const Routes = () =>
             }, 'utils')
           }}
           path="utils"
+        />
+        <Route
+          getComponent={(_nextState, callback) => {
+            require.ensure([], require => {
+              callback(null, require('./css/Components'))
+            }, 'components')
+          }}
+          path="components"
         />
       </Route>
