+++ src/styles/_core.scss
@@ -1,0 +1,2 @@
+@import "core/layout";
+@import "core/breakpoints";
