+++ src/styles/_tools.scss
@@ -1,4 +1,2 @@
 @import "tools/maps";
-@import "tools/mediaquery";
 @import "tools/color";
-@import "tools/layout";
