+++ src/styles/_utilities.scss
@@ -1,10 +1,12 @@
 @import "utilities/mixins";
 @import "utilities/background";
 @import "utilities/layout";
-@import "utilities/typography";
+@import "utilities/text";
 @import "utilities/tables";
 @import "utilities/float";
 @import "utilities/toggle";
 @import "utilities/display";
 @import "utilities/alignment";
 @import "utilities/spacing";
+@import "utilities/typeset";
+@import "utilities/theme";
