+++ src/styles/portrait.scss
@@ -6,6 +6,7 @@
 @import 'tools';
 @import 'base';
 
+@import 'core';
 @import 'custom';
 @import 'bootstrap';
